--- src/com/topicos/aluno/AlunoDAO.java
@@ -1,4 +1,0 @@
-package com.topicos.aluno;
-
-public class AlunoDAO {
-}
