+++ src/com/topicos/app/App.java
@@ -8,7 +8,6 @@ package com.topicos.app;
 
 
 import com.topicos.bancoDeDados.Banco;
-import com.topicos.login.Login;
 import com.topicos.telas.frames.principalFrame.Principal;
 
 public class App {
