+++ src/com/topicos/cadastro/Idiomas.java
@@ -3,7 +3,7 @@ Criado por: Herick Maciel Maia
 Data: 25/03/18
  */
 
-package com.topicos.comum.cadastro;
+package com.topicos.cadastro;
 
 public class Idiomas {
