+++ src/com/topicos/cadastro/aluno/AlunoDAO.java
@@ -1,0 +1,12 @@
+package com.topicos.cadastro.aluno;
+
+public class AlunoDAO extends Aluno{
+
+    private AlunoDAO(){}
+
+    public static void salvarAluno(Aluno aluno) {
+
+    }
+
+
+}
